requirements.txt
aiofiles==23.2.1
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.2
anyio==4.9.0
APScheduler==3.11.0
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.1.31
charset-normalizer==3.3.2
frozenlist==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==5.4.0
magic-filter==1.0.12
multidict==6.2.0
pillow==11.2.1
propcache==0.3.0
pypng==0.20220715.0
python-docx==1.1.0
python-dotenv==1.1.0
python-telegram-bot==22.1
pytz==2024.1
qrcode==7.4.2
reportlab==4.0.7
sniffio==1.3.1
SQLAlchemy==2.0.27
typing_extensions==4.14.0
tzdata==2024.1
tzlocal==5.3.1
urllib3==2.2.2
yarl==1.18.3
